added password input to create account modal
